Add guard clause.

+++ dataloading/finetuning_from_bucket/mosaic_ml_streaming/finetune_mosaic.py
@@ -122,7 +122,8 @@ if __name__ == "__main__":
         (ModelArguments, DataTrainingArguments, TrainingArguments, DDPArguments)
     )
     model_args, data_args, training_args, ddp_args = parser.parse_args_into_dataclasses()
-    shutil.rmtree(data_args.local_cache_path)
+    if os.path.isdir(data_args.local_cache_path):
+        shutil.rmtree(data_args.local_cache_path)
     os.makedirs(data_args.local_cache_path)
     logging.info(f"Created {data_args.local_cache_path}")
     main(model_args, data_args, training_args, ddp_args)
